Restructured directories to distinguish between containers and components
