+++ src/components/IconLink.js
@@ -3,16 +3,18 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import IconButton from 'material-ui/IconButton';
 
-const IconLink = ({ to, tooltip, icon }) => (
-    <Link to={to}>
+const IconLink = ({ to, tooltip, icon }) => {
+    const iconButton = (
         <IconButton iconClassName="material-icons" tooltip={tooltip}>
             {icon}
         </IconButton>
-    </Link>
-);
+    );
+
+    return to ? <Link to={to}>{iconButton}</Link> : iconButton;
+};
 
 IconLink.propTypes = {
-    to: PropTypes.string.isRequired,
+    to: PropTypes.string,
     tooltip: PropTypes.string.isRequired,
     icon: PropTypes.string.isRequired,
 };
