+++ src/containers/GroupList/index.js
@@ -1,0 +1,2 @@
+import GroupList from './GroupList';
+export default GroupList;
