+++ src/containers/RoleDetails.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { USER_ROLE, DETAILS } from '../../constants/entityTypes';
-import DetailSummary from '../DetailSummary';
+import { USER_ROLE, DETAILS } from '../constants/entityTypes';
+import DetailSummary from '../components/DetailSummary';
 import { connect } from 'react-redux';
-import { USER_ROLE_DETAILS } from '../../constants/detailFieldConfigs';
-import { getItem } from '../../actions';
+import { USER_ROLE_DETAILS } from '../constants/detailFieldConfigs';
+import { getItem } from '../actions';
 
 class RoleDetails extends Component {
     componentWillMount() {
