+++ src/containers/RoleForm/RoleForm.js
@@ -5,59 +5,43 @@ import i18next from 'i18next';
 import { Field, reduxForm } from 'redux-form';
 import Heading from 'd2-ui/lib/headings/Heading.component';
 import RaisedButton from 'material-ui/RaisedButton';
-import { navigateTo } from '../../../utils';
-import { clearItem, showSnackbar, getList } from '../../../actions';
-import { NAME, USERS, MANAGED_GROUPS, FIELDS } from './config';
-import { USER_GROUP } from '../../../constants/entityTypes';
+import { navigateTo } from '../../utils';
+import { clearItem, showSnackbar, getList } from '../../actions';
+import { NAME, DESCRIPTION, AUTHORITIES, FIELDS } from './config';
+import { USER_ROLE } from '../../constants/entityTypes';
 import validate from './validate';
-import asyncValidateUniqueness from '../../../utils/asyncValidateUniqueness';
-import api from '../../../api';
-import { renderSearchableGroupEditor } from '../../../utils/fieldRenderers';
-import { asArray } from '../../../utils';
+import asyncValidateUniqueness from '../../utils/asyncValidateUniqueness';
 
-class GroupForm extends Component {
-    saveGroup = (values, _, props) => {
-        const { group, showSnackbar, clearItem, getList } = this.props;
+class RoleForm extends Component {
+    saveRole = (values, _, props) => {
+        const { role, showSnackbar, clearItem, getList } = this.props;
+        role[NAME] = values[NAME];
+        role[DESCRIPTION] = values[DESCRIPTION];
+        role[AUTHORITIES] = values[AUTHORITIES].map(value => ({ id: value }));
 
-        group[NAME] = values[NAME];
-        group[USERS] = values[USERS].map(({ id }) => ({ id }));
-        group[MANAGED_GROUPS] = values[MANAGED_GROUPS].map(({ id }) => ({ id }));
-
-        group
+        role
             .save()
             .then(() => {
-                const msg = i18next.t('User group saved successfully');
+                const msg = i18next.t('User role saved successfully');
                 showSnackbar({ message: msg });
                 clearItem();
-                getList(USER_GROUP);
+                getList(USER_ROLE);
                 this.backToList();
             })
             .catch(error => {
-                const msg = i18next.t('There was a problem saving the user group.');
+                const msg = i18next.t('There was a problem saving the user role.');
                 showSnackbar({ message: msg });
             });
     };
 
     backToList = () => {
-        navigateTo('/user-groups');
+        navigateTo('/user-roles');
     };
 
     renderFields() {
-        const { group } = this.props;
         return FIELDS.map(fieldConfig => {
             const { name, fieldRenderer, label, isRequiredField, ...conf } = fieldConfig;
-            const suffix = isRequiredField ? ' *' : '';
-            const labelText = i18next.t(label) + suffix;
-
-            if (fieldRenderer === renderSearchableGroupEditor) {
-                conf.availableItemsQuery = api[conf.availableItemsQuery];
-                conf.availableItemsLabel = i18next.t(conf.availableItemsLabel);
-                conf.assignedItemsLabel = i18next.t(conf.assignedItemsLabel);
-                if (isRequiredField) {
-                    conf.assignedItemsLabel += ' *';
-                }
-                conf.initialValues = fieldConfig.initialItemsSelector(group);
-            }
+            let labelText = i18next.t(label);
 
             return (
                 <Field
@@ -71,13 +55,13 @@ class GroupForm extends Component {
         });
     }
 
-    render() {
+    render = () => {
         const { handleSubmit, asyncValidating, pristine, valid } = this.props;
         const disableSubmit = Boolean(asyncValidating || pristine || !valid);
         return (
             <main>
                 <Heading level={2}>{i18next.t('Details')}</Heading>
-                <form onSubmit={handleSubmit(this.saveGroup)}>
+                <form onSubmit={handleSubmit(this.saveRole)}>
                     {this.renderFields()}
                     <div style={{ marginTop: '2rem' }}>
                         <RaisedButton
@@ -95,39 +79,39 @@ class GroupForm extends Component {
                 </form>
             </main>
         );
-    }
+    };
 }
 
-GroupForm.propTypes = {
+RoleForm.propTypes = {
     showSnackbar: PropTypes.func.isRequired,
     clearItem: PropTypes.func.isRequired,
     getList: PropTypes.func.isRequired,
     handleSubmit: PropTypes.func.isRequired,
     initialValues: PropTypes.object.isRequired,
-    group: PropTypes.object.isRequired,
+    role: PropTypes.object.isRequired,
     asyncValidating: PropTypes.oneOfType([PropTypes.bool, PropTypes.string]).isRequired,
     pristine: PropTypes.bool.isRequired,
     valid: PropTypes.bool.isRequired,
 };
 
 const mapStateToProps = state => ({
-    group: state.currentItem,
+    role: state.currentItem,
     initialValues: {
         [NAME]: state.currentItem[NAME],
-        [USERS]: asArray(state.currentItem[USERS]),
-        [MANAGED_GROUPS]: asArray(state.currentItem[MANAGED_GROUPS]),
+        [DESCRIPTION]: state.currentItem[DESCRIPTION],
+        [AUTHORITIES]: state.currentItem[AUTHORITIES],
     },
 });
 
-const ReduxFormWrappedGroupForm = reduxForm({
-    form: 'groupForm',
+const ReduxFormWrappedRoleForm = reduxForm({
+    form: 'roleForm',
     validate,
     asyncValidate: asyncValidateUniqueness,
     asyncBlurFields: [NAME],
-})(GroupForm);
+})(RoleForm);
 
 export default connect(mapStateToProps, {
     clearItem,
     showSnackbar,
     getList,
-})(ReduxFormWrappedGroupForm);
+})(ReduxFormWrappedRoleForm);
