+++ src/containers/RoleList/index.js
@@ -1,0 +1,2 @@
+import RoleList from './RoleList';
+export default RoleList;
