+++ src/containers/UserList/index.js
@@ -1,0 +1,2 @@
+import UserList from './UserList';
+export default UserList;
