+++ src/containers/UserProfile.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { USER, DETAILS } from '../../constants/entityTypes';
-import DetailSummary from '../DetailSummary';
+import { USER, DETAILS } from '../constants/entityTypes';
+import DetailSummary from '../components/DetailSummary';
 import { connect } from 'react-redux';
-import { USER_PROFILE } from '../../constants/detailFieldConfigs';
-import { getItem } from '../../actions';
+import { USER_PROFILE } from '../constants/detailFieldConfigs';
+import { getItem } from '../actions';
 
 class UserProfile extends Component {
     componentWillMount() {
